+++ packages/react-app-locale-utils/src/index.ts
@@ -1,5 +1,8 @@
 export { default as defaultLocale } from './defaultLocale.js';
+export { default as displayNames } from './displayNames.js';
 export { default as locales } from './locales.js';
 export { default as useDateFnsLocale } from './useDateFnsLocale.js';
+export { default as useDisplayName } from './useDisplayName.js';
 export { default as useLocale } from './useLocale.js';
+export { default as useMaterialUILocale } from './useMaterialUILocale.js';
 export { default as useMUILocale } from './useMUILocale.js';