+++ packages/react-app-locale-utils/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.0
+
+- **Breaking:** Optimize `useDateFnsLocale`, requires `babel-plugin-codegen`
+- Add `displayNames`
+- Add `useDisplayName`
+- Add `useMaterialUILocale` for `@material-ui/core@4`
+
 # 0.1.1
 
 - Add missing dependency